feat(boardmenu): implement showCard

+++ src/main/java/dio/kanban/dto/BoardColumnDetailsDto.java
@@ -15,6 +15,6 @@ public class BoardColumnDetailsDto {
 
     private BoardColumnKindEnum kind;
 
-    private int cardsAmount;
+    private long cardsAmount;
 
 }
